Create Python_033.py
33.	WAP to create an email collector. 
Input: A raw string as input from the user. 
Output: Create a file Allgmail.txt and store all the gmails present in the string in that file. You need to print on the screen all the gmails collected in the file AllGmail.txt.
Hint: You need to use re (regular expression) module of Python.